# src/api/openapi.yaml
openapi: 3.1.1
info:
  title: Nexxus API
  version: 0.0.11
  summary: HTTP surface of a Nexxus deployment.
  description: |
    Nexxus is a config-driven, real-time backend-as-a-service. This document
    describes the **framework** contract: the routes every deployment exposes
    and the envelope every response carries.

    ### Application-defined shapes

    An application's data models are *runtime configuration*, not source code.
    A deployment serving a fitness app and one serving a chat app expose the
    same `/model/*` routes with entirely different payloads. Where the shape is
    the application's to declare, the schema here says so — see `AppModel` and
    `UserDetails`.

    The same applies to `/auth/{strategy}`: which strategies exist is per
    deployment (`api.auth.availableStrategies`) and per application
    (`auth.strategies`).

    ### Multi-tenancy

    One deployment serves many applications. Almost every request carries an
    `nxx-app-id` header naming which one. It is not a secret and can ship in
    client code.

    ### Authentication

    A session is two tokens, issued by `/user/register`, `/auth/{strategy}` and
    `/device/register`:

    - **`token`** — the access token, sent as `Authorization: Bearer`. It always
      names a **device**, and additionally names a **user** when the application
      has authentication enabled. It expires after the application's
      `session.jwtExpiresIn` (one hour by default). There is no device header —
...

...
servers:
...
paths:
...
  /auth/{strategy}:
    post:
      tags: [ Authentication ]
      summary: Authenticate with a configured strategy
      description: |
        Authenticates against one of the application's configured strategies
        and returns a session.

        - `local` expects `username` and `password` in the body, and returns
          the session directly.
        - Redirect strategies (e.g. `google`) respond with a `302` to the
          provider and complete at `/auth/{strategy}/callback`.

        Which strategies exist is per deployment and per application. `local`
        and `google` ship with the framework; others are resolved from
        `node_modules` by package name. A `404` means this application has not
        configured that strategy.
      security: []
      parameters:
        - $ref: '#/components/parameters/AppId'
        - name: strategy
          in: path
          required: true
          description: Strategy name, e.g. `local` or `google`.
          schema:
            type: string
          example: local
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: Local strategy only.
                password:
                  type: string
                  format: password
                  description: Local strategy only.
                userType:
                  type: string
                  description: |
                    Redirect strategies only — the user type an account is
                    created with if this is a first sign-in.
                  default: default
                device:
                  $ref: '#/components/schemas/DeviceHint'
      responses:
        '200':
          $ref: '#/components/responses/Session'
        '302':
          description: Redirect-based strategies send the caller to the provider.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: |
            Authentication failed. The message does not distinguish an unknown
            username from a wrong password.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
                  - type: object
                    properties:
                      error:
                        const: UserAuthenticationFailedException
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/Unavailable'
  /auth/{strategy}/callback:
    get:
      tags: [ Authentication ]
      summary: Redirect-flow callback
      description: |
        Completes a redirect flow and returns a session. The provider sends
        the caller here; a client does not call it directly. Registered only for
        strategies that declare `requiresCallback`.

        No `nxx-app-id` header — the application is identified by `state`, which
        must come back exactly as issued and can only be redeemed once.
      security: []
      parameters:
        - name: strategy
          in: path
          required: true
          schema:
            type: string
          example: google
        - name: state
          in: query
          required: true
          description: The signed state issued when the flow started.
          schema:
            type: string
        - name: code
          in: query
          description: Provider authorization code.
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/Session'
        '400':
          description: |
            The `state` parameter is missing, malformed, unsigned, or has
            already been redeemed.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
                  - type: object
                    properties:
                      error:
                        const: InvalidParametersException
        '401':
          description: |
            The provider declined, or returned an account whose email address it
            has not verified. An unverified provider email is refused.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
                  - type: object
                    properties:
                      error:
                        const: UserAuthenticationFailedException
        '404':
          $ref: '#/components/responses/NotFound'
  /auth/refresh:
    post:
      tags: [ Authentication ]
      summary: Refresh a session
      description: |
        Exchanges a refresh token for a new access token and a new refresh token.
        Works the same for every application, including those without
        authentication. No `Authorization` header is needed.

        Each refresh replaces the refresh token: store the new one and discard the
        old. The old one is still accepted for a few seconds; presenting it after
        that ends the session.

        The new access token carries the user as currently stored, so it reflects
        changes made through `PUT /user`.
      security: []
      parameters:
        - $ref: '#/components/parameters/AppId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ refreshToken ]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: The next pair of tokens.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: The new access token.
                  refreshToken:
                    type: string
                    description: The new refresh token. The one sent is replaced.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: |
            The refresh token can't be exchanged: malformed, unknown, expired,
            already replaced, or ended by logout. Sign in again.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
                  - type: object
                    properties:
                      error:
                        const: InvalidRefreshTokenException
        '404':
          $ref: '#/components/responses/AppNotFound'
  /auth/logout:
    post:
      tags: [ Authentication ]
      summary: End a session
      description: |
        Ends the session a refresh token belongs to, and closes the device's live
        transport connection if it has one. Works the same for every application,
        including those without authentication.

        Answers `200` to any request that carries a refresh token, whether or not
        a session ended.

        The access token stays valid until it expires.
      security: []
      parameters:
        - $ref: '#/components/parameters/AppId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ refreshToken ]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: The request was processed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/AppNotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The access token. Issued by `/user/register`, `/auth/{strategy}`,
        `/device/register` and `/auth/refresh`. Names a device, and a user where
        the application has authentication.
  parameters:
    AppId:
      name: nxx-app-id
      in: header
      required: true
      description: |
        Which application this request is for. Not a secret; it can ship in
        client code.
      schema:
        type: string
...
  schemas:
    Error:
      type: object
      description: |
        Every failure uses this shape. `error` is a stable identifier to branch
        on; `message` is for humans and may change.
      required: [ error, message ]
      properties:
        error:
          type: string
          description: Identifies the failure.
        message:
          type: string
        stack:
          type: string
          description: Present only when the deployment runs with `NODE_ENV=dev`.
    DeviceHint:
      type: object
      description: |
        What a client can say about the device it is calling from. An `id` that
        does not resolve to a device this caller owns is ignored and a new
        device is issued; it is not an error.
      properties:
        id:
          type: string
          description: The device id from a previous session.
        name:
          type: string
          description: Used only when a new device is created.
...
  responses:
    Session:
      description: |
        The session's two tokens, the device they are bound to, and the user they
        name.
      content:
        application/json:
          schema:
            type: object
            properties:
              token:
                type: string
                description: The access token.
              refreshToken:
                type: string
                description: Exchanged at `POST /auth/refresh` for the next pair.
              device:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
              user:
                type: object
                properties:
                  id:
                    type: string
                  username:
                    type: string
...
    BadRequest:
      description: The request is malformed or a parameter is invalid.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  error:
                    enum:
                      - InvalidParametersException
                      - InvalidAuthMethodException
...
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  error:
                    const: NotFoundException
    AppNotFound:
      description: No application matches the `nxx-app-id` header.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  error:
                    const: ApplicationNotFoundException
...
    Unavailable:
      description: |
        An upstream service is currently disconnected. Retry.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  error:
                    const: ServiceUnavailableException
security:
  - bearerAuth: []